Moved threads into api folder
