+++ app/app/src/main/java/com/example/nightout/api/ImageRetrievalThread.java
@@ -1,4 +1,4 @@
-package com.example.nightout;
+package com.example.nightout.api;
 
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
